tests/wpt/mozilla/tests/mozilla/empty_clientrect.html

<html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(function() {
  var rect = window.document.head.getBoundingClientRect();
  var rects = window.document.head.getClientRects();
  assert_true(rect instanceof DOMRect, "Should be DOMRect");
  assert_equals(rect.top, 0);
  assert_equals(rect.bottom, 0);
  assert_equals(rect.left, 0);
  assert_equals(rect.right, 0);
  assert_equals(rect.width, 0);
  assert_equals(rect.height, 0);
  assert_equals(rects.length, 0);
});
</script>
</head>
</html>
